// src/types/Company.ts
export interface Company {
  companyName: string;
  sector: string;
  currentPrice: string;
  highPrice: string;
  lowPrice: string;
  ceo: string;
  employees: string;
  marketCap: string;
  dividends: string;
  sdgs: string[];
  logoBackground: string;
  logoText: string;
  bullish: boolean;
  id: string;
}